* More warnings killed.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8662 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/codecs/flac/libs/libFLACHelper/FLACHeaderTweaker.cpp
@@ -125,7 +125,7 @@ bool FLACHeaderTweaker::createNewHeaderList() {
 			//It's the comment packet.
 			//debugLog<<"Found a comment packet..."<<endl;
 			locFoundComment = true;
-			locCommentNo = i;
+			locCommentNo = (int)i;
 			mNewHeaderList.push_back(mOldHeaderList[i]->clone());
 		}
 	}
@@ -138,6 +138,7 @@ bool FLACHeaderTweaker::createNewHeaderList() {
 
 	for (size_t i = 2; i < mOldHeaderList.size(); i++) {
 	
+		//**** WARNING ::: Leave this unless you check it !
 		if (i != locCommentNo) {
 			//debugLog<<"Adding another ehader..."<<endl;
 			//If it's not the comment packet we already added, put it in the list.
